requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.35.1
pyserial==3.5
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
joblib==1.3.2
